requirements.txt
glcontext==3.0.0
moderngl==5.11.1
numpy==2.1.0
pygame==2.6.0
pygame-ce==2.5.1
pygame_gui==0.6.12
PyGLM==2.7.1
python-i18n==0.3.9
PyWavefront==1.3.3
